# Avoid allocating table on stack.