@nigeltao 2016-02-13
I still have some significant concerns about this PR in its entirety. One point is that it does at least three things: it changes the hash table value type from int to int32, it uses a sync.Pool to avoid mallocs and subsequent GC, and uses an 'encoder.cur' value to avoid having to reset the hash table to zeroes. Sure each change is a small amount of code, but they are separable changes.

As I said on https://go-review.googlesource.com/#/c/19336/:

"""
If it turns out that things A, B and D are all relatively simple code changes and gets you +9%, and C is very complicated and gets you +1%, I would recommend rejecting C even if that +1% was statistically significant. Raw MB/s throughput is not the only criteria in which the Go standard library code is judged. As this CL is packaged now, though, we either have to accept all four or reject all four.

Accordingly, my recommendation would be to break this CL up into a smaller ones... As always, smaller yet still self-contained CLs are easier to review, and easier to manage (e.g. roll back if necessary).
"""

Another point is that, as I said above a few days ago, I'm still not 100% convinced that the e.cur mechanism is correct when it overflows. You check for t < 0, and then for t >= s (via the offset variable), but I think that it is still possible to have 0 <= t && t < s yet the hash table entry being from a previous sync.Pool use. Sure, it would be very, very rare for that to happen, but also, very, very hard to reproduce bugs are the hardest ones to debug.

Related to that concern is code like "*p = int32(s + sadd)" where s (an int, being in the range 0 <= s && s < len(src)) doesn't necessarily fit into an int32 and you can silently overflow.

Sure, what you've done may actually end up being 100% safe, but I'd need some convincing, and I'd need that proof to be written as a comment, and it'd have to be something longer than "If t >= s this will be negative, when converted to a uint this will always be > maxOffset."

As I also said above a few days ago, I'd like to see additional tests, especially around the wrap point. As it is, there are no new tests in this PR.

In a previous PR, #19, you have already expressed some frustration that you are wasting your time with my code review of your pull requests.

You have also mentioned that your goals aren't necessarily the same as my goals, e.g. byte-for-byte compatibility with the C++ encoding algorithm.

Accordingly, I suspect that it's not worth us both spending time iterating on this PR until I'm satisfied with it. I'm happy to continue the conversation if I'm wrong and you do want to spend more time on it, but otherwise, I'll incorporate some of your suggestions as separate commits, such as cc71ae7.