@nigeltao 2016-02-14
I'm probably flogging a dead horse, in the metaphorical sense, but a "byte-for-byte compatibility with the C++ encoding algorithm" test would have at least detected the bug fixed by d1d908a2 for free, without requiring an explicit "I wonder if these benchmark improvements are too good to be true" investigation.