@klauspost 2016-02-14
Good point. Here is a test:
```Go
// TestEncodeNoiseThenRepeats encodes a 8MB block for which the first half is
// very incompressible and the second half is very compressible. The encoded
// form's length should be closer to 50% of the original length than 100%.
func TestEncodeNoiseThenRepeatsBig(t *testing.T) {
	const origLen = 8 << 20
	src := make([]byte, origLen)
	rng := rand.New(rand.NewSource(1))
	firstHalf, secondHalf := src[:origLen/2], src[origLen/2:]
	for i := range firstHalf {
		firstHalf[i] = uint8(rng.Intn(256))
	}
	for i := range secondHalf {
		secondHalf[i] = uint8(i >> 8)
	}
	dst := Encode(nil, src)
	if got, want := len(dst), origLen*3/4; got >= want {
		t.Fatalf("got %d encoded bytes, want less than %d", got, want)
	}
}
```
The official Snappy has a maximum block size of 64k (not 1GB), which prevents this bug from showing up, but also eliminates all matches across blocks. My own solution is `s += 1 + (((s-lit)>>4) & 127)`, which limits skips to 127 bytes, but keeps the progressive step-up.

So, you are left with the choice, if you want C++ compatibility and make compression worse(~1.5%), or keep the IMO better compressor?

btw, if you want compatibility in that aspect, you've made it very easy, it's just `const maxInternalEncodeSrcLen = (1 << 16)`.