@nigeltao 2016-04-03
I went for C++ compatibility. As of 8939696c, it should now output byte-for-byte identical output as the C++ snappy encoder.

In an earlier comment, I noted that this PR does at least three things:
1. it changes the hash table value type from int to int32,
2. it uses a sync.Pool to avoid mallocs and subsequent GC, and
3. uses an 'encoder.cur' value to avoid having to reset the hash table to zeroes.

Point 1 was done separately in an earlier commit (cc71ae7c). Having reached C++ compatibility, I am curious about new numbers for points 2 and 3, but I would still like to see them as separate changes, and I would still like some extra tests if possible and especially extra commentary for point 3.

If you are happy to keep working on this PR to re-base and address those points, I'm willing to keep this PR open.

If you have no shortage of other things to do with your time, I'm otherwise happy to work on those points myself over the next week or two.