@nigeltao 2016-04-11
I re-based this patch (and committed it on a separate branch). I might have screwed up somewhere, but I'm not convinced that points 2 and 3 are a net win on the benchmarks any more, but they still make the code more complicated.

See https://github.com/golang/snappy/commit/fdf5d5de7f4e9fb3760d9421a87402c249a05326