@nigeltao 2017-02-15
The snappy codebase has changed a lot since this PR, so I'm closing this as obsolete. Thanks anyway for all of the optimization discussions.